Make the login experience more smooth
Instead of copying the access token from the browser, gonedrive will now
get it automatically.
It uses the authcode - accesstoken workflow.
First the authcode will be received by opening an own server which
closes as soon as microsoft redirects to it.
After getting the auth code, a POST request will be send to a token URL
which contains the token as body response.