Introduce new command 'secret'
With 'secret' gonedrive is able to upload secrets to the cloud. With a
given name and the value of a credential, it will upload it to an
encrypted secret file. For more details run 'gonedrive secret --help'.

Files which implement the endpoint of azure are now all under
'service/azure'.
Also an interface is defined which other cloud endpoints should
implement, e.g. Google Drive.